requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
litellm==1.48.8
pydantic==2.8.2
httpx==0.27.0